chore: Display item list in the home page
