+++ client/src/common/layout/head/index.js
@@ -1,0 +1,12 @@
+import Head from 'next/head'
+
+export default function HeadComponent () {
+  return (
+    <Head>
+      <title>React Hooks Playground</title>
+      <meta name="description" content="React Hooks Playground" />
+      <meta name="viewport" content="width=device-width, initial-scale=1" />
+      <link rel="icon" href="/favicon.ico" />
+    </Head>
+  )
+}
