+++ client/src/pages/usesyncexternalstore/index.js
@@ -1,0 +1,15 @@
+import HeadComponent from '@/common/layout/head'
+
+function useSyncExternalStoreContainer () {
+  return (
+    <>
+      <HeadComponent />
+
+      <h2>
+        useSyncExternalStoreContainer
+      </h2>
+    </>
+  )
+}
+
+export default useSyncExternalStoreContainer
