+++ client/src/styles/globals.css
@@ -1,3 +1,7 @@
+html, body {
+  font-family: Arial, Helvetica, sans-serif;
+}
+
 :root {
   --max-width: 1100px;
   --border-radius: 12px;
@@ -87,12 +91,6 @@ body {
 
 body {
   color: rgb(var(--foreground-rgb));
-  background: linear-gradient(
-      to bottom,
-      transparent,
-      rgb(var(--background-end-rgb))
-    )
-    rgb(var(--background-start-rgb));
 }
 
 a {
